py3 Changes in __del__ for gitpython
gitpython in py2 relied upon the behavior of __del__ to release locks on
config_writer objects. In py3, because __del__ may happen much later
than when an object drops out of scope, we must explicitly release locks
either with the .release() method or by using the config writer as a
context manager.

Change-Id: Ic09eb7af6757eefbb0f8c86cb1bf3e8ca5f007b9