Fix passing labels to Gerrit when they are not defined in All-Projects
This is a port of Ib591997141b68708bee0b291bb88a93cd3294301 to zuulv3
as it somehow got lost.

Turns out that the shortcut --verified or --code-review only work when
these labels are defined globally for all projects [1]. The other syntax
with a --label works fine even when e.g. the Verified label is being
used just for a subset of projects.

[1] https://gerrit-review.googlesource.com/#/c/51800/1/Documentation/cmd-review.txt

Change-Id: Ia9d3e3280995163ae2caee8ed25c7e987fcb00a4