Add override-branch property to job repos
When a job specifies what repos it requires, allow it to also
specify a branch to check out specifically for that project.
This is similar to the project_override_branch option in zuul-cloner
which we have used to indicate that regardless of the presence of,
say, a stable/icehouse branch, we should always check out the
master branch of tempest in a particular job.

Change-Id: Ie4a5350b48eb02cadfbaf4a242770ede9a89fd1b