Add support for defining groups in nodesets
If users are expected to be able to use ansible content written for
production, it is important to be able to define arbitrary groups of
nodes in their inventory. For instance, a playbook to deploy OpenStack
may want a groups called controller, compute, ceph-osd and ceph-monitor,
but a job to test that playbook may want three nodes, one called
compute, one called controller1 and one called controller2. For the test
job, I would want to put controller1 in the ceph-osd group and
controller1 and controller2 in the ceph-monitor group.

nodepool does not need to know anything about these - they are just
logical names the user is describing to make it into the inventory.

There are currently no tests of the inventory we're writing out. The
next patch adds a test to ensure that inventories are written out
properly.

Change-Id: I5555c86ffa96e6a43df5e46302f4e76840372999