Use the executor cached repos more often
Currently, we clone the job's working repos from the repo cache that
the executor maintains.  However, some playbook and role repos do
not use that cache.

To facilitate more operations using the merger class, add the concept
of a repo cache to it, and use that in all job related git repo operations
on the executor.

Change-Id: I747b6602540458506e1f6d480a95b80c6543c5a8