Enable test_one_branch cloner test
This alters test_cloner to test the behavior of the executor, which,
in v3, performs the git repo preparation that is performed by the
cloner in v2.  These tests represent all of the ways we expect zuul
to be used with git repos in different kinds of jobs, so the test
is altered as little as possible.

A future change will rename the test file and class.

Change-Id: I4dc9aeb325e38c54b4150bcc8ac672096e2cadbc