Fix missing source connections on merger
Change I453754052b75de3d4fac266dc73921c7ce9c075f had the effect that
no source connection was configured in the merger. This can be fixed
by changing the type check from issubclass to isinstance. Further also
the zuul and timer drivers don't need to be configured by the merger.

This also adds a test case for that.

Change-Id: If437123c3e8a0d63c5f852619e214cefa0892ce3