Add LoggerAdapter for executor jobs
Create a logger adapter for executor jobs that ensures that every
log entry has the job uuid associated with it.  Pass this into
the associated mergers as well, since they are also performing
interesting tasks in parallel on the executor.

The uuid is also added to the extras dict so that it's available
to logging systems which can handle extra data.

Change-Id: I28e0bcd0f030361659e5a72c162d549c1f0d6acb