Remove test_periodic_update
This test verified that before running a periodic job, the merger
updated its repos so that zuul-cloner would fetch the latest
merged changes.

Much of this process is not relevant anymore, however, to the
extent that we want to verify that periodic jobs do actually get
the latest copies of repositories, the test_post* jobs cover
this sufficiently.  They exercise the same mechanism of updating
and cloning repos with the latest information for non-change
related pipelines.

Change-Id: I002244c84d96b71179cc0726c79d8180ca2ca526