Strip double-timestamps from output log
ansible-playbook uses basicConfig at the top of the command, which
essentially takes over everything if you send things to log_path.

Instead, go ahead and log things ourselves so that our log lines can all
look consistent (and so that we can easily html-ify) We pass in the
location via an environment variable since we're not using the log_dir
option in ansible.cfg.

As of this patch, there will be less output. A follow up will add back
the missing callback plugin bits.

Change-Id: Ic3ff57bba7a3a23dc5d0055e8e9888f24641f7d5