Add known_hosts to bubblewrap jobir.work_root
Because we already have the known_hosts for all nodes we are planning
to use for a job, it makes sense to prime bubblewrap with this
information.  This is specifically needed if using the verify_host
field in synchronize, because it relies on the ssh client knowning
have to read the default known_hosts file.

Change-Id: Ifdb3ac1eb7443beacb9277b5749d773b0c6aa4ad
Signed-off-by: Paul Belanger <pabelanger@redhat.com>