Read layout from BuildSet in first merge scenario
When you are running the first test of a project then the layout doesn't
contain the project and so can't report. It seems like this used to be
able to be pulled from the QueueItem.layout but this doesn't seem to be
set any more. If we get it from the BuildSet this works so we should
allow it to be sourced from there.

Change-Id: Icf972d1c5eb351e7800c05064f94b86513df660b
Signed-off-by: Jamie Lennox <jamielennox@gmail.com>