Use worker_name for job cancellation and remove manager
The word "manager" in zuul primarily means "PipelineManager" and is
confusing here. We also send back "worker_name" which is intended to be
the unique name for the worker.

For now, send back hostname for the worker_name. In the future, we can
update this to be overridable per-executor in the config.

While we're at it, remove no-longer valid references to other worker
data from the docs.

Change-Id: Ibe5cf7295f133c9dc48162b40ae1f625a19643dc