Remove job_name_in_report option
This option was to maintain backwards compatability with the report
format used by the jenkins gerrit trigger plugin.  We no longer
need to worry about that.  With the addition of various reporters
which may need to report information in different formats, we
should probably either move this to the gerrit driver, or make it
customizable.

For now, let's go ahead and remove the option and use the behavior
we would have if it were set to true, which has been the case for
OpenStack infra's and all of our third-party ci users for some
time.

Change-Id: I8f040e03b7074cf59eced6fca38e6b221051d93e