Implement Depends-On for github
This adds the capability for one github project pr to depend on another
project's PR, by marking it so in the pull request body. This means we
need to be able to react to editing of the pull request body to see if a
dependency was created or removed.

Reverse searching for a change that needs the current change is
implemented via a github search of (open) pull request bodies.

Various bits of how change objects are created and cached needed to be
fixed to support a change being added while adding another change.

Also put a try around the webhook event handling, so that a traceback,
such as when a dependency loop is detected, can be gracefully logged
instead of the whole thing shutting down.

Change-Id: Ia32e6cc70b163f7e32e74d88f0a1f9b4e3255320
Story: 2000774
Task: 4686