Add linebreak to generated passwd/group file
For running in bwrap the /etc/passwd and /etc/group files are
generated on the fly to only show the executing user. This needs to
add a linebreak at the end. Otherwise ssh (as well as getent) cannot
read the file. In case of ssh this results in the error 'No user
exists for uid x'.

Change-Id: I0e75dd423f2ffb93da1de4dfc064ff22991f1793