Create a new logger for gerrit IO
Gerrit io traffic used to be in a separate logger due to being in
a different class, but with subsequent reorganizations, we now
have all gerrit traffic being logged with the connection.

Create a new zuul.GerritConnection.io logger so that that massive
amount of data doesn't end up in normal debug logs and can be
filtered separately.

Change-Id: Ie54802455a8679b1618f057c1e0b8dfd17fe2b19