Augment --keep-jobdirs with an IPC
It's difficult to start zuul in the normal way via with an extra
argument for debugging.  Make this an IPC accessible toggle.

The argument is kept for the time being because it's used by the
unit tests, and it's the only way to specify the value on startup,
but we should replace it later with a config file setting.

Change-Id: Ia59a9383fcf90a00e1475977629b7d71d3a40cb0