Don't automatically mount user home in executor
We're starting to treat the work directory as a substitute home
directory (we put .ssh/ into it, for example), and we set $HOME
to that directory.  Complete this process by updating our bwrap
passwd entry to point to that as the home directory and stop
mounting the real home dir.

Change-Id: I0fdb1913634d3902cac58112c5d683f12675c6f7