Always run post playbooks
Post playbooks get us things like logs.  Including logs from
pre-playbooks.

Generally, we want to retry the job if a pre-playbook fails.  However,
if it keeps failing, we run into the RETRY_LIMIT and stop trying.  In
that case, we would like to see the logs from the failed job.  To make
that happen, always run all of the post-playbooks for a job, even if
the pre-playbooks have failed.

Change-Id: I8457e5a89a3cc34c351f1f861847c5fbe231e28e