Handle item loops in zuul_stream
Items in loops get their own callback events, so we need to handle them.

Additionally, the final item and the event finished event come out of
sequence - so we need to save a note in the event finished event and
then emit it after the following item.

Finally, we'd have a todo list item from ansiblefest to copy not
directly modify the result objects we get because they're shared. Turns
out this is immediately necessary with items, because popping
stdout_lines in one item totally borks the further items (whoops)

Change-Id: Ia70239f240276b10ad09d4bda57310ebcbdb1f4d