Improve debugging at start of job
It would be really helpful if the name of the job is prominently
displayed in the debug log when the executor starts running it.

Right now the ref is not as useful as it should be, but that should
change when we finish moving away from zuul_ref.

The unique id is no longer required since we automatically prepend
it to every log entry.

Change-Id: Ic005771705c548d363868f70bab868564e24591c