Fix test_bubblewrap_leak
We've seen this test return a false positive, presumably because
the sleep process is not immediately reaped by the bubblewrap
pid 1 process.  To compensate, sleep for much longer (60s), and
give bwrap 30 seconds to actually terminate the sleep.  As long
as we verify the sleep ends early, this test should be valid.

Change-Id: I828cc0b9c34a4ebc34d8432a776f2cd6e642b399