Handle GitHub comment reviews carefully
A comment review should not mask a previously provided approval or
changes_requested review from a given user. The GitHub UX follows the
same logic. If an approval or a changes_requested review was already
provided from a user, that user can issue future comment reviews and it
will not change their vote. Our algorithm needed to be updated to handle
this scenario, and a test case was added to validate it.

Change-Id: Iac10544a1f24362c27220a44edbdb988b7caa989