Add support for marking module results uninteresting
Using python to collect data for later display is useful, as it allows
us to control what's output and what isn't. That's defeated by the
default behavior of logging all of the result parameters. Add a flag we
can set in result dicts that will cause us to discard the parameters
from the log on success. On failure we don't want to do this because on
failure all the debugging we can get is great.

Change-Id: Ic2295c7a8be30144fe47c5bd385e827bfd8095f2