Fail execution if a role repo is a bare collection of roles
We generally expect a repo to either be a single bare role, or
to have a "roles/" directory which is a collection of roles.

We previously also supported the repo itself being a collection
of roles at the root of the repo, but that's not widely used
and is potentially confusing.  Especially since we want to make
a followup change to implicitly treat some repos as roles repos.

Stop supporting that for now.  If folks want to support this,
we might consider doing so only if there is a meta/main.ya?ml
file in the repo (as Galaxy accidentally supports that, though
it is not an explicit contract they guarantee).

Change-Id: Ie70e9fae4da57d4aefa01442a9f60526163b27c0