Filter and print results more comprehensively
Don't emit results when no_log is set. That's impotant.

If the result is a message and a rc, emit the message - since
that's the important message for the user.

If messages are multi-line, split them and show them multi-line.

If the results have an "invocation" in them, suppress it.

This will still result in synchronize calls potentially showing a lot of
output - but it should be in a readable format. If output from a
synchronize call is not desired or two large, it can be marked no_log.

Change-Id: I90cf5de8eb4479773aee97b5590ac125c06dc00e